requirements.txt
alembic==0.9.8
aniso8601==3.0.0
astroid==1.6.1
blinker==1.4
bottle==0.12.13
cffi==1.11.5
click==6.7
coverage==4.5.1
cryptography==1.7.2
flake8==3.5.0
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.7.1
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.14.2
pluggy==0.6.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.5.3
pyasn1==0.4.2
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
PyJWT==1.6.0
pylint==1.8.2
python-dateutil==2.7.0
python-dotenv==0.7.1
python-editor==1.0.3
pytz==2018.3
six==1.11.0
SQLAlchemy==1.2.4
tox==3.0.0
virtualenv==15.2.0
Werkzeug==0.14.1
wrapt==1.10.11
